% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{aggregate_expand}
\alias{aggregate_expand}
\title{Oblicz wartość funkcji dla podzbiorów i rozwiń do oryginalnej
długości.}
\usage{
aggregate_expand(dv, f, fun = mean)
}
\arguments{
\item{dv}{Zmienna agregowana.}

\item{f}{Czynnik grupujący.}

\item{fun}{Funkcja agregująca (ma zwracać skalar).}
}
\value{
Wektor zagregowanych wartości rozwinięty do długości length(dv).
}
\description{
Oblicz wartość funkcji dla podzbiorów i rozwiń do oryginalnej
długości.
}

