% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{binomial.pred.ci}
\alias{binomial.pred.ci}
\title{Przybliżone przedziały ufności dla proporcji}
\usage{
binomial.pred.ci(fits, n, alpha = 0.05)
}
\arguments{
\item{fits}{Wektor wartości oczekiwanych (prawdopodobieństw)}

\item{n}{Wektor wielkości prób}

\item{alpha}{1 - prawdopodobieństwo pokrycia}
}
\description{
Rozkład normalny może być kiepskim przybliżeniem przedziałów
ufności (w wikipedii jest odnośnik do artykułu opisującego kiedy ta
reguła nie działa). Generalnie proporcja nie może być zbyt blisko 0
ani 1. Istnieje wiele innych metod liczenia interwałów, wiki sporo
o tym mówi.
}

