% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bp}
\alias{bp}
\title{MCMC with bipartite graph}
\usage{
bp(out = "run1", whole = "T2D.whole", part = "T2D.part",
  edge = "T2D.edge", alpha = 0.05, beta = 0.2, pi = 0.01,
  nburn = 10000, ngen = 100000, sub = 1000, penalty = 2,
  initial = c("inactive", "random", "high"))
}
\arguments{
\item{out}{Root name for output files}

\item{whole}{Input file with whole node info}

\item{part}{Input file with part node info}

\item{edge}{Input file with edge info}

\item{alpha}{Parameter alpha (0 < alpha < beta < 1)}

\item{beta}{Parameter alpha (0 < alpha < beta < 1)}

\item{pi}{Parameter pi (0 < pi < 1)}

\item{nburn}{Number of burn-in generations}

\item{ngen}{Number of sample generations}

\item{sub}{Subsample rate for burn-in and sample files}

\item{penalty}{Penalty per illegal node to loglikelihood}

\item{initial}{Initial state (see Details)}
}
\description{
MCMC with bipartite graph
}
\details{
The \code{initial} argument can take one of three values:
\code{"inactive"} - all whole nodes inactive; \code{"random"} - all
whole nodes active with probability \code{pi}, no illegal nodes; or
\code{"high"} - all nodes with proportion of connected part nodes
with response equal to 1 above 0.4 are active, no illegal nodes.
}
\keyword{models}

