% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bp_jhs}
\alias{bp_jhs}
\title{Blood Pressure - John Schwenck}
\format{
An object of class \code{data.frame} with 222 rows and 15 columns.
}
\source{
<google.com>
}
\usage{
bp_jhs
}
\description{
Blood Pressure - John Schwenck
}
\keyword{datasets}
