% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_range.R
\name{bp_range}
\alias{bp_range}
\title{Blood Pressure Range}
\usage{
bp_range(data, inc_date = 0)
}
\arguments{
\item{data}{Required dataframe with SBP and DBP columns corresponding to
Systolic and Diastolic BP. This dataframe should come from \code{data_process}}

\item{inc_date}{Optional argument that specifies whether or not to break down
range by date. Default value is \code{inc_date = 0} indicating that date is
omitted from grouping. This is typically omitted for ABPM data as dates tend
to overlap due to recordings during sleep.}
}
\value{
A tibble with SBP_max, SBP_min, SBP_range, DBP_max, DBP_min, DBP_range
and any additional optional columns included in data such as \code{ID}, \code{VISIT},
\code{WAKE}, and \code{DATE}
}
\description{
Calculates the range (max - min) of both SBP and DBP values with the option
to specify date as an additional level of granularity
}
\examples{
# Load hypnos_data
data(hypnos_data)
data(bp_jhs)

# Process hypnos_data
data1 <- process_data(hypnos_data, sbp = "SYST", dbp = "DIAST", bp_datetime = "date.time",
id = "id", wake = "wake", visit = "visit", hr = "hr", pp ="pp", map = "map", rpp = "rpp")
# Process bp_jhs data
data2 <- process_data(bp_jhs, sbp = "Sys.mmHg.", dbp = "Dias.mmHg.", bp_datetime = "DateTime",
hr = "Pulse.bpm.")

# Calculate BP range
bp_range(data1)
bp_range(data2, inc_date = 1)
}
