% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_report.R
\name{bp_report}
\alias{bp_report}
\title{Blood Pressure Report}
\usage{
bp_report(
  data,
  filename = "bp_report.pdf",
  width = 12,
  height = 8.53,
  filetype = "pdf",
  loc = NULL,
  save_report = 1
)
}
\arguments{
\item{data}{Test}

\item{filename}{Test}

\item{width}{Test}

\item{height}{Test}

\item{filetype}{Test}

\item{loc}{Test}

\item{save_report}{Test}
}
\value{
Test
}
\description{
Test
}
\examples{
# Test
}
