% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_sleep_metrics.R
\name{bp_sleep_metrics}
\alias{bp_sleep_metrics}
\title{Blood Pressure Sleep Metrics}
\usage{
bp_sleep_metrics(data, subj = NULL)
}
\arguments{
\item{data}{User-supplied data set containing blood pressure data. Must
contain a Systolic blood pressure (\code{SBP}), Diastolic blood pressure (\code{DBP})
as well as \code{ID}, \code{WAKE}, and \code{DATE_TIME} columns.}

\item{subj}{Optional argument. Allows the user to specify and subset specific subjects
from the \code{ID} column of the supplied data set. The \code{subj} argument can be a single
value or a vector of elements. The input type should be character, but the function will
comply with integers so long as they are all present in the \code{ID} column of the data.}
}
\value{
The function outputs a list containing: (1) a list of count data related
to how many instances were observed during night / during day, (2) a list of all
summary statistics for SBP, (3) a list of all summary statistics for DBP, and (4)
a list of all medical literature metrics.
}
\description{
The \code{bp_sleep_metrics} function serves to calculate sleep-dependent metrics
from relevant medical literature.
}
\examples{
hypnos_proc <- process_data(bp_hypnos,
                           sbp = "syst",
                           dbp = "DIAST",
                           date_time = "date.time",
                           id = "id",
                           wake = "wake",
                           visit = "visit",
                           hr = "hr",
                           map = "map",
                           rpp = "rpp",
                           pp = "pp")

bp_sleep_metrics(hypnos_proc)

}
