% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_stages.R
\name{bp_stages}
\alias{bp_stages}
\title{Alternative Blood Pressure Stages}
\usage{
bp_stages(
  data,
  sbp,
  dbp,
  inc_low = TRUE,
  inc_crisis = TRUE,
  data_screen = TRUE
)
}
\arguments{
\item{data}{User-supplied data}

\item{sbp}{column name corresponding to systolic blood pressure (SBP)}

\item{dbp}{column name corresponding to diastolic blood pressure (DBP)}

\item{inc_low}{A TRUE / FALSE indicator of whether or not to include the "Low" (Hypotension)
category to the scatter plot. The range for Hypotension is set from a minimum of 25 for DBP or 80
for SBP, or the corresponding minimum value for either category from the data until 60 for DBP and
100 for SBP.}

\item{inc_crisis}{A TRUE / FALSE indicator of whether or not to include the Hypertensive "Crisis"
category to the scatter plot. The range for crisis is any value above 180 for SBP or above 120 for
DBP.}

\item{data_screen}{Default to TRUE. data_screens for extreme values in the data for both \code{SBP} and \code{DBP}
according to Omboni, et al (1995) paper - Calculation of Trough:Peak Ratio of Antihypertensive Treatment
from Ambulatory Blood Pressure: Methodological Aspects}
}
\value{
A dataframe with additional columns corresponding to the stages of high blood pressure and the
supplementary SBP / DBP categories
}
\description{
Adds SBP_Category and DBP_Category columns to supplied dataframe.
}
\details{
Supplied dataframe must adhere to the unified format using the \code{process_data} function.
}
\examples{
# Load bp_hypnos
data(bp_hypnos)

bp_stages(bp_hypnos, sbp = "syst", dbp = "diast")


# Load bp_jhs data
data(bp_jhs)

bp_stages(bp_jhs, sbp = "sys.mmhg.", dbp = "dias.mmhg.")

}
