% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{df_from_files}
\alias{df_from_files}
\title{Przeszukaj (pod)katalogi i sklej ramki wczytane z plików pasujących
do podanego wzorca.}
\usage{
df_from_files(test.fun, verbose = TRUE, header = FALSE, sep = "\\t",
  fun = function(fname) {     res = read.table(fname, header = header, sep =
  sep, ...)     res }, ...)
}
\description{
Przeszukaj (pod)katalogi i sklej ramki wczytane z plików pasujących
do podanego wzorca.
}

