% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dip_calc.R
\name{dip_calc}
\alias{dip_calc}
\title{Nocturnal Blood Pressure Dipping Calculation}
\usage{
dip_calc(data, sleep_int = NULL, dip_thresh = 0.1)
}
\arguments{
\item{data}{User-supplied data set that must contain SBP, DBP, and either DATE_TIME or WAKE columns in order to distinguish between sleep and awake}

\item{sleep_int}{(optional) User-supplied sleep interval to indicate start and end time of the interval of interest. For example, sleep_int = c(22, 5) indicates a sleep period from 10pm - 5am. Must only contain 2 values and must be in 24-hour format. If no sleep_int supplied and no WAKE column provided in data set, default interval is from 11pm - 6am according to (paper)}

\item{dip_thresh}{(optional) Default threshold to classify as either a dip, non-dip, or reverse pattern. The default threshold is 0.10 (i.e. 10\% dip)}
}
\value{

}
\description{
Calculate the percent and average drop (or potentially reverse) in nocturnal blood pressure.
This function is typically used with abpm data; in the event of non-abpm data, the sleep_int argument must be utilized to specify what times correspond to sleep vs awake.

NOTE: The DATE_TIME column MUST be supplied in the data in order to perform calculations
}
