% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{do.files}
\alias{do.files}
\title{Rekurencyjne operowanie na plikach/scierzkach.}
\usage{
do.files(test = function(x) TRUE, fun = function(x, result) message(x),
  dir = ".", result = NULL)
}
\description{
Dla kazdego pliku, jesli spelnia zadany test to wykonaj na nim
zadana operacje. Jesli jest katalogiem to wejdz i rekurencyjnie rób
dalej swoje. dir to katalog początkowy.
}

