% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{do.files.test}
\alias{do.files.test}
\title{Test funkcji \code{do.files} - wypisanie wszystkich plikóww w
podkatalogach.}
\usage{
do.files.test(dir)
}
\description{
Test funkcji \code{do.files} - wypisanie wszystkich plikóww w
podkatalogach.
}

