% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_visuals_tables.R
\name{dow_tod_plots}
\alias{dow_tod_plots}
\title{Blood Pressure Tables in Graphical Format}
\usage{
dow_tod_plots(data)
}
\arguments{
\item{data}{A processed dataframe resulting from the \code{process_data} function that
contains the \code{SBP}, \code{DBP}, \code{DAY_OF_WEEK}, \code{Time_of_Day}, \code{SBP_Category},
and \code{DBP_Category} columns.}
}
\value{
A list of four `gtables` that correspond to the tables for Day of Week and
Time of Day broken down by both \code{SBP} and \code{DBP}.
}
\description{
Blood Pressure Tables in Graphical Format
}
\examples{

data("hypnos_data")
hyp_proc <- process_data(hypnos_data,
                         sbp = "syst",
                         dbp = "DIAST",
                         bp_datetime = "date.time",
                         id = "id",
                         wake = "wake",
                         visit = "visit",
                         hr = "hr",
                         map = "map",
                         rpp = "rpp",
                         pp = "pp",
                         ToD_int = c(5, 13, 18, 23))

rm(hypnos_data)

dow_tod_plots_tmp <- dow_tod_plots(hyp_proc)
grid::grid.draw(
   gridExtra::grid.arrange(dow_tod_plots_tmp[[1]], dow_tod_plots_tmp[[2]],
                           dow_tod_plots_tmp[[3]],dow_tod_plots_tmp[[4]],
                           nrow = 2, ncol = 2)
                           )
}
