% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessers.R
\name{find_guessers}
\alias{find_guessers}
\title{Dopasowuje model służący do wykrywania osób wykonujących zadanie
przypadkowo}
\usage{
find_guessers(score, n, guessing_prob = 0.5,
  n.chains = parallel::detectCores() - 1, sample = 20000, ...)
}
\arguments{
\item{score}{Wektor sum poprawnych odpowiedzi}

\item{n}{Wektor liczby prób}

\item{guessing_prob}{= .5, zakładane prawdopodobieństwo przypadkowego trafienia}

\item{...}{Pozostałe argumenty dla funkcji run.jags}
}
\value{
Lista z elementami fit i pred, gdzie fit to obiekt zwracany przez run.jags a pred to ramka z predykcjami posteriorycznymi
}
\description{
Dopasowuje model służący do wykrywania osób wykonujących zadanie
przypadkowo
}

