% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypnos_data}
\alias{hypnos_data}
\title{HYPNOS Study - 5 Subject Sample}
\format{
An object of class \code{data.frame} with 250 rows and 14 columns.
}
\usage{
hypnos_data
}
\description{
Dexcom G4 CGM measurements for 5 subjects with Type II diabetes. These data are part of a larger study sample that consisted of patients with Type 2 diabetes recruited from the general community. To be eligible, patients with Type 2 diabetes, not using insulin therapy and with a glycosylated hemoglobin (HbA$_1c$) value at least 6.5
ABPM measurements
}
\details{
description of variables
}
\keyword{datasets}
