% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire.R
\name{ll.compute}
\alias{ll.compute}
\title{Oblicza wynik surowy dla kwestionariusza Liebovitz'a - skala lęk}
\usage{
ll.compute(x, ...)
}
\description{
Oblicza wynik surowy dla kwestionariusza Liebovitz'a - skala lęk
}

