% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{lmer.sig}
\alias{lmer.sig}
\title{Poziomy istotności dla mieszanego modelu liniowego}
\usage{
lmer.sig(m, return = FALSE, short = FALSE, show = TRUE)
}
\arguments{
\item{m}{Model dopasowany za pomocą funkcji lmer.}

\item{return}{Czy zwracać tabelkę regresji jako obiekt.}

\item{short}{Czy drukować tabelkę w wersji skróconej.}

\item{show}{Czy drukować tabelkę.}
}
\value{
Zwraca tabelkę z istotnościami, jeżeli return ==
TRUE. Drukuje tabelkę z istotnościami na wyjściu, jeżeli show ==
TRUE.
}
\description{
Poziomy istotności dla mieszanego modelu liniowego
}

