% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{lms}
\alias{lms}
\title{Funkcja zwracająca ładną prostą tabelkę z gwiazdkami dla modelu
liniowego}
\usage{
lms(..., digits = 3, return.table = FALSE)
}
\description{
Funkcja zwracająca ładną prostą tabelkę z gwiazdkami dla modelu
liniowego
}

