% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire.R
\name{lu.compute}
\alias{lu.compute}
\title{Oblicza wynik surowy dla kwestionariusza Liebovitz'a - skala unikanie}
\usage{
lu.compute(x, ...)
}
\description{
Oblicza wynik surowy dla kwestionariusza Liebovitz'a - skala unikanie
}

