% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitsandpieces.R
\name{p2star}
\alias{p2star}
\title{Poziom istotności zamieniony na gwiazdki}
\usage{
p2star(x)
}
\arguments{
\item{x}{Wektor poziomów istotności}
}
\value{
Wektor tekstowy zawierający 0 lub więcej gwiazdek
}
\description{
Poziom istotności zamieniony na gwiazdki
}

