% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire.R
\name{panas20.compute}
\alias{panas20.compute}
\title{Oblicza wynik surowy dla kwestionariusza PANAS-C 20}
\usage{
panas20.compute(x, means = TRUE, ...)
}
\description{
Oblicza wynik surowy dla kwestionariusza PANAS-C 20
}

