% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire.R
\name{panas20s.compute}
\alias{panas20s.compute}
\title{Oblicza wynik surowy dla kwestionariusza PANAS-S 20}
\usage{
panas20s.compute(x, means = TRUE, ...)
}
\description{
Oblicza wynik surowy dla kwestionariusza PANAS-S 20
}

