% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{path_check}
\alias{path_check}
\title{Compatibility Check for \code{path} Argument}
\usage{
path_check(path)
}
\arguments{
\item{path}{A user-supplied string corresponding to the path location
where the export is to be saved}
}
\value{
A formatted string corresponding to the proper binding of
directory and base naming conventions. If user's path is invalid,
\code{path_check} will throw an error.
}
\description{
Compatibility Check for \code{path} Argument
}
\examples{
\dontrun{
path = "~dir/abpm/"
path_check(path) # drops the trailing slash
}
}
