% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{rmatch}
\alias{rmatch}
\title{Zwraca wektor logiczny odpowiadający literałom pasującym do wzorca}
\usage{
rmatch(pattern, vector)
}
\description{
Zwraca wektor logiczny odpowiadający literałom pasującym do wzorca
}

