% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{robust.mixed}
\alias{robust.mixed}
\title{Dopasowuje odporny mieszany model liniowy lub logistyczny}
\usage{
robust.mixed(fixed, random, d, y_nu = 4, y_sigma = 1.548435, ranef_nu = 4,
  beta_sigma = 20, chains = parallel::detectCores(), pars = NULL,
  family = "binomial", ...)
}
\arguments{
\item{fixed}{Formuła modelu dla efektów ustalonych.}

\item{random}{Formuła modelu dla efektów losowych, postaci g ~ x1 +
x2 ...}

\item{d}{Zbiór danych.}

\item{y_nu}{Liczba stopni swobody rozkładu t modelującego zmienną
zależną.}

\item{y_sigma}{(Uwzględniane tylko dla modelu logistycznego)
Odchylenie standardowe rozkładu t modelującego zmienną
zależną. To jest wartość arbitralna - domyślnie przyjęto
wartość dającą współczynniki porównywalne do zwykłej regresji
logistycznej.}

\item{ranef_nu}{Liczba stopni swobody rozkładu t modelującego
rozkład efektów losowych.}

\item{beta_sigma}{Odchylenie standardowe prioru dla efektów
ustalonych. UWAGA Jeżeli predyktory mają bardzo małe odchylenie
standardowe, albo oczekiwane prawdopodobieństwa dla punktów
przecięcia są skrajne, domyślna wartość może być
nieodpowiednia.}

\item{chains}{Liczba równoległych próbników - domyślnie liczba
rdzeni.}

\item{pars}{Wektor tekstowy nazw parametrów, dla których chcemy
próbki, poza beta, ranef_sigma i C. Może zawierać 'y_new'
(posterior predykcyjny) i 'ranef' (efekty losowe).}
}
\value{
Wynik próbkowania zwrócony przez funkcję stan z dodanym
    atrybutem fixed.effects, zawierającym nazwy kolumn macierzy
    modelu.
}
\description{
Implementacja modelu jest oparta na publikacji Sorensen'a,
Hohenstain'a i Vasishth'a (2015): Bayesian linear mixed models
using Stan: A tutorial for psychologists, linguists, and cognitive
scientists. Warto pamiętać, że prior dla efektów ustalonych nie
jest płaski (domyślnie beta_sigma = sd prioru normalnego == 20).
}

