% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt.R
\name{roc_points}
\alias{roc_points}
\title{Oblicza proporcje trafień i fałszywych alarmów dla poszczególnych ocen}
\usage{
roc_points(stim, dec, rating, max.rating = max(rating), f = 1)
}
\arguments{
\item{stim}{Wektor binarny reprezentujący bodziec.}

\item{dec}{Wektor binarny reprezentujący decyzję.}

\item{rating}{Wektor reprezentujący oceny na skali 1-max.rating.}

\item{max.rating}{Maksymalna wartość oceny. Domyślnie max(rating).}

\item{f}{Czynnik, który ma oddzielać podzbiory danych do obliczania punktów ROC.}
}
\value{
Ramka z kolumnami f, resp (przekodowane oceny), hi i fa.
}
\description{
Oblicza proporcje trafień i fałszywych alarmów dla poszczególnych ocen
}

