% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{stage_check}
\alias{stage_check}
\title{Compatibility Checks for User-Supplied SBP/DBP Stages Vector}
\usage{
stage_check(sbp_stages = NULL, dbp_stages = NULL)
}
\arguments{
\item{sbp_stages}{Checks whether the user supplied \code{sbp_stages_alt} function argument is valid or not}

\item{dbp_stages}{Checks whether the user supplied \code{dbp_stages_alt} function argument is valid or not}
}
\value{
A list containing two vectors that each correspond to the SBP/DBP stages for use in further processing.
If the function returns no error, the supplied vectors are valid. Function arguments are coerced to integer.
}
\description{
Used in \code{process_data} function
}
