% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{subject_subset_check}
\alias{subject_subset_check}
\title{Compatibility Check for User-Supplied Subject Subset Vector}
\usage{
subject_subset_check(data, subj = NULL)
}
\arguments{
\item{data}{Input data within original function}

\item{subj}{A vector corresponding to a selected subset of subjects to
refine the supplied data with}
}
\value{
Logical indicating whether the subj vector passes the checks (\code{TRUE}),
or not (\code{FALSE})
}
\description{
Compatibility Check for User-Supplied Subject Subset Vector
}
