% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{t2d}
\alias{t2d}
\title{Type 2 diabetes data}
\format{A list with three components:
\code{whole}: list of character strings with the names of the "whole nodes";
\code{part}: a vector of 0's and 1's; names are the names of the "part nodes";
\code{edges}: a matrix of with two columns of character strings, with each
row defining an edge between a "whole node" and a "part node".}
\source{
Bret Larget
}
\usage{
data(t2d)
}
\description{
Not entirely sure what these data are...
}
\examples{
data(t2d)
}
\keyword{datasets}

