% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_sleep_metrics.R
\name{bp_sleep_metrics}
\alias{bp_sleep_metrics}
\title{Blood Pressure Sleep Metrics}
\usage{
bp_sleep_metrics(data, subj = NULL)
}
\arguments{
\item{data}{User-supplied data set containing blood pressure data. Must
contain a Systolic blood pressure (\code{SBP}), Diastolic blood pressure (\code{DBP})
as well as \code{ID}, \code{WAKE}, and \code{DATE_TIME} columns.}

\item{subj}{Optional argument. Allows the user to specify and subset specific subjects
from the \code{ID} column of the supplied data set. The \code{subj} argument can be a single
value or a vector of elements. The input type should be character, but the function will
comply with integers so long as they are all present in the \code{ID} column of the data.}
}
\value{
The function outputs a list containing 4 tibble objects corresponding to the following tables:
\item{[[1]]}{Counts of how many BP measurements were observed overall (\code{N_total}), during sleep (\code{N_sleep}), and during wake (\code{N_awake}) for each subject ID and grouping variable}
\item{[[2]]}{Summary statistics for systolic BP measurements (SBP):}
\item{[[3]]}{Summary statistics for diastolic BP measurements (DBP), same as for SBP}
\item{[[4]]}{BP metrics associated with sleep: dip percentage (\code{dip_calc}), nocturnal fall (\code{noc_fall}), morning blood pressure surge (mbps) for sleep-through (\code{ST_mbps}) and pre-wake (\code{PW_mbps})}
}
\description{
The \code{bp_sleep_metrics} function serves to calculate sleep-dependent metrics
from relevant medical literature.
}
\examples{
hypnos_proc <- process_data(bp_hypnos,
                           sbp = "syst",
                           dbp = "DIAST",
                           date_time = "date.time",
                           id = "id",
                           wake = "wake",
                           visit = "visit",
                           hr = "hr",
                           map = "map",
                           rpp = "rpp",
                           pp = "pp")

bp_sleep_metrics(hypnos_proc)

}
