% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{bp_stages}
\alias{bp_stages}
\title{Alternative Blood Pressure Stages}
\usage{
bp_stages(
  data,
  sbp,
  dbp,
  sbp_stages_alt = NULL,
  dbp_stages_alt = NULL,
  screen = FALSE
)
}
\arguments{
\item{data}{User-supplied data}

\item{sbp}{column name corresponding to systolic blood pressure (SBP)}

\item{dbp}{column name corresponding to diastolic blood pressure (DBP)}

\item{sbp_stages_alt}{(Optional) Determines the lower and upper limits for each stage
of systolic blood pressure (\code{SBP}). If supplied, the input must be a vector
containing 7 integers that correspond to the limits of each stage. The default vector would be
given by \code{c(80, 100, 120, 130, 140, 180, 200)} where:
\itemize{

   \item Low - default: < 100 (specifically 80 - 100)
   \item Normal - default: 100 - 120
   \item Elevated - default: 120 - 130
   \item Stage 1 - default: 130 - 140
   \item Stage 2 - default: 140 - 180
   \item Crisis - default: > 180 (specifically 180 - 200)

}}

\item{dbp_stages_alt}{(Optional) Determines the lower and upper limits for each stage
of diastolic blood pressure (\code{DBP}). If supplied, the input must be a vector
containing 7 integers that correspond to the limits of each stage. The default vector would be
given by \code{c(25, 60, 80, 85, 90, 120, 140)} where:
\itemize{

   \item Low - default: < 60 (specifically 25 - 60)
   \item Normal - default: 60 - 80
   \item Elevated - default: 80 - 85
   \item Stage 1 - default: 85 - 90
   \item Stage 2 - default: 90 - 120
   \item Crisis - default: > 120 (specifically 120 - 140)

}}

\item{screen}{Default to TRUE. Screens for extreme values in the data for both \code{SBP} and \code{DBP}
according to Omboni, et al (1995) paper - Calculation of Trough:Peak Ratio of Antihypertensive Treatment
from Ambulatory Blood Pressure: Methodological Aspects}
}
\value{
A dataframe with two additional columns corresponding to \code{SBP_Category} and \code{DBP_Category}
}
\description{
Adds SBP_Category and DBP_Category columns to supplied dataframe.
}
\details{
Supplied dataframe must adhere to the unified format using the \code{process_data} function.
}
\examples{

# Load bp_hypnos
data("bp_hypnos")

bp_stages(bp_hypnos,
            sbp = "syst",
            dbp = "diast",
            sbp_stages_alt = c(80, 100, 120, 130, 140, 170, 200),
            dbp_stages_alt = c(25, 60, 80, 85, 90, 120, 140)
            )


# Load bp_jhs data
data("bp_jhs")

bp_stages(bp_jhs,
            sbp = "sys.mmhg.",
            dbp = "dias.mmhg.",
            sbp_stages_alt = c(80, 100, 120, 130, 160, 170, 200),
            )

}
