% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypnos_data}
\alias{hypnos_data}
\title{HYPNOS Study - 5 Subject Sample}
\format{
An object of class \code{data.frame} with 250 rows and 14 columns.
}
\usage{
hypnos_data
}
\description{
HYPNOS Study - 5 Subject Sample
}
\keyword{datasets}
