% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessers.R
\name{plot_guessers}
\alias{plot_guessers}
\title{Wykres dopasowania modelu wykrywającego osoby wykonujące zadanie na poziomie przypadku}
\usage{
plot_guessers(fit)
}
\arguments{
\item{fit}{Obiekt zwracany przez find_guessers}
}
\description{
Wykres dopasowania modelu wykrywającego osoby wykonujące zadanie na poziomie przypadku
}

