% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{process_data}
\alias{process_data}
\title{Data Pre-Processor}
\usage{
process_data(
  data,
  sbp = NULL,
  dbp = NULL,
  bp_datetime = NULL,
  id = NULL,
  wake = NULL,
  visit = NULL,
  hr = NULL,
  pp = NULL,
  map = NULL,
  rpp = NULL
)
}
\arguments{
\item{data}{User-supplied dataset containing blood pressure data. Must
contain data for Systolic blood pressure and Diastolic blood pressure at a
minimum.}

\item{sbp}{Required column name string corresponding to Systolic Blood
Pressure (mmHg)}

\item{dbp}{Required column name string corresponding to Diastolic Blood
Pressure (mmHg)}

\item{bp_datetime}{Optional column name string corresponding to Date/Time,
but HIGHLY recommended to supply if available.}

\item{id}{Optional column name string corresponding to subject ID. Typically
needed for data corresponding to more than one subject.}

\item{wake}{Optional column name string corresponding to sleep status. A
WAKE value of 1 indicates that the subject is awake and 0 implies asleep.}

\item{visit}{Optional column name string corresponding to Visit number}

\item{hr}{Optional column name string corresponding to Heart Rate (bpm)}

\item{pp}{Optional column name string corresponding to Pulse Pressure
(SBP - DBP). If not supplied, it will be calculated automatically.}

\item{map}{Optional column name string corresponding to Mean Arterial
Pressure}

\item{rpp}{Optional column name string corresponding to Rate Pulse
Pressure (SBP * HR). If not supplied, but HR column available, then
RPP will be calculated automatically.}
}
\value{
A processed dataframe object
}
\description{
A helper function to assist in pre-processing the user-supplied
input data for use with other functions.
Typically, this function will process the data and be stored as another
dataframe. See Vignette for further details.
}
