% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{robust_mixed}
\alias{robust_mixed}
\title{Dopasowuje odporny mieszany model liniowy lub logistyczny}
\usage{
robust_mixed(fixed, random, d, n = NULL, y_nu = 4, y_sigma = 1.548435,
  beta_sigma = NULL, ranef_nu = 4, chains = parallel::detectCores() - 1,
  pars = NULL, family = "binomial", return_stanfit = F, ...)
}
\arguments{
\item{fixed}{Formuła modelu dla efektów ustalonych.}

\item{random}{Formuła modelu dla efektów losowych, postaci g ~ x1 +
x2 ...}

\item{d}{Zbiór danych.}

\item{n}{Liczba obserwacji na punkt danych (robit)}

\item{y_nu}{Liczba stopni swobody rozkładu t modelującego zmienną
zależną.}

\item{y_sigma}{(Uwzględniane tylko dla modelu logistycznego)
Odchylenie standardowe rozkładu t modelującego zmienną
zależną. To jest wartość arbitralna - domyślnie przyjęto
wartość dającą współczynniki porównywalne do zwykłej regresji
logistycznej.}

\item{beta_sigma}{Odchylenie standardowe prioru dla efektów
ustalonych. UWAGA Jeżeli predyktory mają bardzo małe odchylenie
standardowe, albo oczekiwane prawdopodobieństwa dla punktów
przecięcia są skrajne, domyślna wartość może być
nieodpowiednia.}

\item{ranef_nu}{Liczba stopni swobody rozkładu t modelującego
rozkład efektów losowych.}

\item{chains}{Liczba równoległych próbników - domyślnie liczba
rdzeni.}

\item{pars}{Wektor tekstowy nazw parametrów, dla których chcemy
próbki, poza beta, ranef_sigma i C. Może zawierać 'y_new'
(posterior predykcyjny) i 'ranef' (efekty losowe).}

\item{family}{(='binomial') Rozkład zmiennej zależnej: binomial
oznacza odporną regresję logistyczną, normal oznacza odporny
model liniowy.}

\item{return_stanfit}{(=F) Czy zwracać obiekt zwracany przez Stana,
czy listę z ramką z próbkami i summary}
}
\value{
Obiekt zwracany przez rstan lub lista złożona z elementów
    s: ramka próbek, summary: podsumowanie wyników STAN'a.
}
\description{
Implementacja modelu jest oparta na publikacji Sorensen'a,
Hohenstain'a i Vasishth'a (2015): Bayesian linear mixed models
using Stan: A tutorial for psychologists, linguists, and cognitive
scientists. Warto pamiętać, że prior dla efektów ustalonych nie
jest płaski (domyślnie beta_sigma = sd prioru normalnego == 20).
}

